BEAM-6693 add a stats module and move approximate uniq functions to stats module
